Remove redundancies and fix related imports
