+++ project/accounts/urls/urls.py
@@ -3,8 +3,6 @@ from django.contrib.auth import views as auth_views
 from accounts.views import (RegisterView, SettingsView, ProfileActivationView, PasswordResetView, PasswordResetDoneView,
                             PasswordResetConfirmView, PasswordResetCompleteView, ProfileSetupView, user_profile)
 
-from threads.views.views import about_view, support_us_view, how_it_works_view, base_view
-
 urlpatterns = [
     path(
         'login/',
@@ -14,6 +12,8 @@ urlpatterns = [
     path('logout/', auth_views.LogoutView.as_view(), name='accounts_logout'),
     path('register/', RegisterView.as_view(), name='accounts_register'),
     path('settings/', SettingsView.as_view(), name='accounts_settings'),
+    path('setup/', ProfileSetupView.as_view(), name="accounts_profile_setup"),
+    path("profile/<str:username>/", user_profile, name="profile"),
     path('activate_account/<uidb64>/<token>/', ProfileActivationView.as_view(), name='accounts_activate'),
     path(
         'accounts/password_reset/',
@@ -35,6 +35,4 @@ urlpatterns = [
         PasswordResetCompleteView.as_view(),
         name='accounts_password_reset_complete',
     ),
-    path('setup/', ProfileSetupView.as_view(), name="accounts_profile_setup"),
-    path("profile/<str:username>/", user_profile, name="profile")
 ]
