+++ project/core/urls.py
@@ -19,8 +19,6 @@ from django.conf.urls.static import static
 from django.contrib import admin
 from django.urls import include, path
 from django.views.generic.base import RedirectView
-from threads import urls as threads
-
 
 from core.router import CiviWikiRouter
