+++ project/threads/urls/api.py
@@ -2,22 +2,8 @@ from django.conf.urls import include
 from django.urls import path
 from rest_framework.routers import DefaultRouter
 
-from threads.api import (
-    create_civi,
-    delete_civi,
-    edit_civi,
-    edit_thread,
-    get_civi,
-    get_thread,
-    rate_civi,
-    upload_civi_image,
-    new_thread,
-    get_civis,
-    get_responses,
-    upload_thread_image,
-)
-
-from threads.views.views import ThreadViewSet, CiviViewSet
+from threads import api
+from threads.views import ThreadViewSet, CiviViewSet
 from categories.api import CategoryViewSet
 from accounts.api import ProfileViewSet
 
@@ -32,20 +18,20 @@ urlpatterns = [
 ]
 
 urlpatterns += [
-    path("thread_data/<int:thread_id>/", get_thread, name="get thread"),
-    path("civi_data/<int:civi_id>/", get_civi, name="get civi"),
-    path("threads/<int:thread_id>/civis", get_civis, name="get civis"),
+    path("thread_data/<int:thread_id>/", api.get_thread, name="get thread"),
+    path("civi_data/<int:civi_id>/", api.get_civi, name="get civi"),
+    path("threads/<int:thread_id>/civis", api.get_civis, name="get civis"),
     path(
         "response_data/<int:thread_id>/<int:civi_id>/",
-        get_responses,
+        api.get_responses,
         name="get responses",
     ),
-    path("new_thread/", new_thread, name="new thread"),
-    path("edit_thread/", edit_thread, name="edit thread"),
-    path("new_civi/", create_civi, name="new civi"),
-    path("rate_civi/", rate_civi, name="rate civi"),
-    path("edit_civi/", edit_civi, name="edit civi"),
-    path("delete_civi/", delete_civi, name="delete civi"),
-    path("upload_images/", upload_civi_image, name="upload images"),
-    path("upload_image/", upload_thread_image, name="upload image"),
+    path("new_thread/", api.new_thread, name="new thread"),
+    path("edit_thread/", api.edit_thread, name="edit thread"),
+    path("new_civi/", api.create_civi, name="new civi"),
+    path("rate_civi/", api.rate_civi, name="rate civi"),
+    path("edit_civi/", api.edit_civi, name="edit civi"),
+    path("delete_civi/", api.delete_civi, name="delete civi"),
+    path("upload_images/", api.upload_civi_image, name="upload images"),
+    path("upload_image/", api.upload_thread_image, name="upload image"),
 ]
